coalesce Executor and BaseExecutor
